Rework request logic for mobile client
