+++ pkg/api/kubernetes/kubernetes.go
@@ -3,7 +3,9 @@ package kubernetes
 import (
 	"encoding/base64"
 	"fmt"
+	"time"
 
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -50,4 +52,16 @@ users:
 	}
 
 	return restClient, clientset, nil
+}
+
+func Request(method, url, body string, timeout int64, clientset *kubernetes.Clientset) ([]byte, error) {
+	if method == "GET" {
+		return clientset.RESTClient().Get().RequestURI(url).Timeout(time.Duration(timeout) * time.Second).DoRaw()
+	} else if method == "DELETE" {
+		return clientset.RESTClient().Delete().RequestURI(url).Timeout(time.Duration(timeout) * time.Second).DoRaw()
+	} else if method == "PATCH" {
+		return clientset.RESTClient().Patch(types.JSONPatchType).RequestURI(url).Body([]byte(body)).Timeout(time.Duration(timeout) * time.Second).DoRaw()
+	}
+
+	return []byte(``), fmt.Errorf("Request method is not implemented")
 }
